+++ xtt/lib/flow/src/flow_text.cpp
@@ -1,5 +1,5 @@
 /* 
- * Proview   $Id: flow_text.cpp,v 1.6 2007-09-25 13:11:00 claes Exp $
+ * Proview   $Id: flow_text.cpp,v 1.7 2007-09-25 16:36:21 claes Exp $
  * Copyright (C) 2005 SSAB Oxelösund AB.
  *
  * This program is free software; you can redistribute it and/or 
@@ -46,7 +46,7 @@ void FlowText::traverse( int x, int y)
 
 void FlowText::print( void *pos, void *node, int highlight)
 {
-  int idx = int( ctx->print_zoom_factor / ctx->base_zoom_factor * 
+  double idx = ( ctx->print_zoom_factor / ctx->base_zoom_factor * 
 		(text_size +4) - 4);
   double size = 8.0 + 6.0/3*idx;
